backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
sqlalchemy==2.0.40
pydantic-settings==2.8.1
python-multipart==0.0.20
openai
